refactor for cleaner pruning proof module
